Update police_radio_transcription.py
commit makes it so that there's no loss of transcription data that was done by the old version. makes sure to splice different time stamps that are encoded by the whispers open.ai module. adds two new columns in the .csv file for the last detected time stamp for the last word transcribed and the length of the audio file being transcribed to help find other potential issues with the script. added in annotations to help clarify what the script does